backend/requirements.txt
# requirements.txt
Django>=4.2
psycopg2-binary
djangorestframework
dj-rest-auth
django-cors-headers
django-allauth